[SHACK-290] Minimum Chef 13 version is 13.10.4 now
Fixes https://github.com/chef/chef-apply/issues/22

Chef suffered from a variety of errors when trying to run with Tarball
support. On Windows the URI file unpacker fails with a GZip error. If we
specify relative paths on Chef 13 it thinks the files are at the root of
the directory.

We fix these by specifying full paths, and setting the minimum chef
versions to ones that include support for:
https://github.com/chef/chef/pull/7522
https://github.com/chef/chef/pull/7523

Signed-off-by: tyler-ball <tball@chef.io>