Modify ChefCore::Text to support adding translations
Since ChefCore::Text is now being exposed for use in other gems
and applications, it must support those gems providing their own
translations in order to supplement the defaults that ship with
chef-core.

This does that via new method Text.add_localization; this can be
invoked from consumers. It accepts the base path to localizations,
and also looked for errors/LANG.yml under that path. THis allows
consumers to use ChefCOre's standardized error formatting
and handling for their own errors.

This PR also fixes a long-standing bug  - when a top-level
R18n key contains only text and no sub-keys, it would raise
an error on access.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>