Make CHEFCCR006 more friendly
In cases where a resource attribute did not exist,
CHEFCCR006 would display the whole class name. With this change
if we can recognize the class name as a Resource, we'll trim
it a bit to make it closer to how it's specified by the operator -
eg "User" instead of "Chef::Resources::User::LinuxUser"

In a future iteration, we might try instantiating the resource and
getting its name.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>