Remove DK dep for chef-core-actions. Temp revert to train.
The DK dependency was a leftover from when DK-dependent actions were
included in chef-core-actions.  There is no need for DK, so it was
removed from the gemspec.

Due to dependency conflicts in chef, we can't move to train-core
until after Inspec 4 is released, so this switches us back to the
full train gem.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>