Partial support for :unix as a supported `base_os`
This currently aliases to linux because the additional supported
behaviors are the same, except for installation. We are not
using InstallAction for other OSes yet, but this support is required for
bootstrap which does not use the InstallAction.

`Actions::Base` has been updated with an overrideable
`supported_base_os_types` method. It defaults to all supported base OSes
[:linux, :unix, :windows].  Prior to running an action
`Base#run` will now check the target host's `base_os` against supported
platforms and raise UnsupportedTargetOS if the base OS is not supported
by the action.

InstallChef was updated to declare it supports only :windows and :linux.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>