Moved telemeter to chef-telemetry gem.
Namespace changed as well since it left ChefCore::

I'm not sure about lib/chef_core/telemeter/patch.rb, since it is
monkey patching another gem we own and that patch causes the tests in
chef-telemetry to fail. Why did we not push it upstream? Why do the
tests here pass without it?

Signed-off-by: Ryan Davis <zenspider@chef.io>